+++ stor/settings.py
@@ -15,6 +15,9 @@ _ENV_VARS = {
         'auth_url': 'OS_AUTH_URL',
         'temp_url_key': 'OS_TEMP_URL_KEY',
         'num_retries': 'OS_NUM_RETRIES'
+    },
+    'dx': {
+        'auth_token': 'DX_AUTH_TOKEN'
     }
 }
 """
@@ -40,7 +43,7 @@ def _get_env_vars():
     """
     Update settings with environment variables, if applicable.
 
-    Currently handles swift credentials.
+    Currently handles swift and dx credentials.
     """
     new_settings = {}
     for section in _ENV_VARS: