+++ stor/swift.py
@@ -54,7 +54,6 @@ from stor import is_swift_path
 from stor import settings
 from stor import utils
 from stor.base import Path
-from stor.obs import OBSFile
 from stor.obs import OBSPath
 from stor.obs import OBSUploadObject
 from stor.posix import PosixPath
@@ -102,7 +101,6 @@ ConditionNotMetError = stor_exceptions.ConditionNotMetError
 ConflictError = stor_exceptions.ConflictError
 UnavailableError = stor_exceptions.UnavailableError
 UnauthorizedError = stor_exceptions.UnauthorizedError
-SwiftFile = OBSFile
 SwiftUploadObject = OBSUploadObject
 
 
@@ -648,7 +646,7 @@ class SwiftPath(OBSPath):
                                  '&'.join(query),
                                  auth_url_parts.fragment))
 
-    def write_object(self, content, **swift_upload_args):
+    def write_object(self, content):
         """Writes an individual object.
 
         Note that this method writes the provided content to a temporary
@@ -660,42 +658,20 @@ class SwiftPath(OBSPath):
 
         Args:
             content (bytes): raw bytes to write to OBS
-            **swift_upload_args: Keyword arguments to pass to
-                `SwiftPath.upload`
         """
         if not isinstance(content, bytes):  # pragma: no cover
-            warnings.warn('future versions of stor will raise a TypeError if content is not bytes')
+            if six.PY2:
+                # bytes/unicode a little confused so allow it
+                warnings.warn('Python 3 stor and a future Python 2 version of stor will raise a'
+                              ' TypeError if content is not bytes')
+            else:
+                raise TypeError('write_object() expects bytes, not text data')
         mode = 'wb' if type(content) == bytes else 'wt'
         with tempfile.NamedTemporaryFile(mode=mode) as fp:
             fp.write(content)
             fp.flush()
             suo = OBSUploadObject(fp.name, object_name=self.resource)
-            return self.upload([suo], **swift_upload_args)
-
-    def open(self, mode='r', encoding=None, swift_upload_options=None):
-        """Opens a `SwiftFile` that can be read or written to.
-
-        For examples of reading and writing opened objects, view
-        `SwiftFile`.
-
-        Args:
-            mode (str): The mode of object IO. Currently supports reading
-                ("r" or "rb") and writing ("w", "wb")
-            encoding (str): text encoding to use. Defaults to
-                ``locale.getpreferredencoding(False)`` (Python 3 only)
-            swift_upload_options (dict): DEPRECATED (use `stor.settings.use()`
-                instead). A dictionary of arguments that will be
-                passed as keyword args to `SwiftPath.upload` if any writes
-                occur on the opened resource.
-
-        Returns:
-            SwiftFile: The swift object.
-
-        Raises:
-            SwiftError: A swift client error occurred.
-        """
-        swift_upload_options = swift_upload_options or {}
-        return SwiftFile(self, mode=mode, encoding=encoding, **swift_upload_options)
+            return self.upload([suo])
 
     @_swift_retry(exceptions=(ConditionNotMetError, UnavailableError))
     def list(self,
@@ -913,16 +889,7 @@ class SwiftPath(OBSPath):
     def download_objects(self,
                          dest,
                          objects):
-        """Downloads a list of objects to a destination folder.
-
-        Note that this method takes a list of complete relative or absolute
-        paths to objects (in contrast to taking a prefix). If any object
-        does not exist, the call will fail with partially downloaded objects
-        residing in the destination path.
-
-        This method retries ``num_retries`` times if swift is unavailable.
-        View `module-level documentation <swiftretry>` for more information
-        about configuring retry logic at the module or method level.
+        """See baseclass for complete method documentation
 
         Args:
             dest (str): The destination folder to download to. The directory
@@ -931,42 +898,12 @@ class SwiftPath(OBSPath):
                 download. The objects can paths relative to the download path
                 or absolute swift paths. Any absolute swift path must be
                 children of the download path
-
         Returns:
             dict: A mapping of all requested ``objs`` to their location on
                 disk
-
         Raises:
             ValueError: This method was called on a path that has no
                 container
-
-        Examples:
-
-            To download a objects to a ``dest/folder`` destination::
-
-                from stor import path
-                p = path('swift://tenant/container/dir/')
-                results = p.download_objects('dest/folder', ['subdir/f1.txt',
-                                                             'subdir/f2.txt'])
-                print results
-                {
-                    'subdir/f1.txt': 'dest/folder/subdir/f1.txt',
-                    'subdir/f2.txt': 'dest/folder/subdir/f2.txt'
-                }
-
-            To download full swift paths relative to a download path::
-
-                from stor import path
-                p = path('swift://tenant/container/dir/')
-                results = p.download_objects('dest/folder', [
-                    'swift://tenant/container/dir/subdir/f1.txt',
-                    'swift://tenant/container/dir/subdir/f2.txt'
-                ])
-                print results
-                {
-                    'swift://tenant/container/dir/subdir/f1.txt': 'dest/folder/subdir/f1.txt',
-                    'swift://tenant/container/dir/subdir/f2.txt': 'dest/folder/subdir/f2.txt'
-                }
         """
         if not self.container:
             raise ValueError('cannot call download_objects on tenant with no container')